backend/requirements.txt
click==7.1.2
datetime
Flask==1.1.2
flask_cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
requests==2.25.1
SQLAlchemy==1.3.22
sqlalchemy-cockroachdb==1.3.2
Werkzeug==1.0.1